Tesis JECG/requirements.txt
statsmodels.api
matplotlib
pandas
numpy